Consolidate categories data actions and reducers
