+++ src/shop-redux-categories.js
@@ -1,0 +1,112 @@
+import { store, installReducers } from './shop-redux-store.js';
+import { findCategory, findCategoryIndex } from './shop-redux-helpers.js';
+
+installReducers({
+  // Init categories
+  _categoriesChanged(state, action) {
+    const categories = action.categories;
+    return {
+      ...state,
+      categories
+    }
+  },
+  // Response from fetch for categories data.
+  _categoryItemsChanged(state, action) {
+    const categories = state.categories.slice(0);
+    const categoryIndex = findCategoryIndex(categories, action.categoryName);
+    categories[categoryIndex] = {...categories[categoryIndex], items: action.data};
+    return {
+      ...state,
+      categories,
+      failure: false
+    };
+  },
+  // Network error (set to true by unsucessful fetch).
+  _failureChanged(state, action) {
+    const failure = action.failure;
+    return {
+      ...state,
+      failure
+    };
+  },
+  // Opposite of navigator.onLine (updated by shop-app).
+  _offlineChanged(state, action) {
+    const offline = action.offline;
+    return {
+      ...state,
+      offline
+    };
+  }
+});
+
+function initCategories() {
+  store.dispatch({
+    type: '_categoriesChanged',
+    categories: [
+      {
+        name: 'mens_outerwear',
+        title: 'Men\'s Outerwear',
+        image: 'images/mens_outerwear.jpg',
+        placeholder: 'data:image/jpeg;base64,/9j/4QAYRXhpZgAASUkqAAgAAAAAAAAAAAAAAP/sABFEdWNreQABAAQAAAAeAAD/7gAOQWRvYmUAZMAAAAAB/9sAhAAQCwsLDAsQDAwQFw8NDxcbFBAQFBsfFxcXFxcfHhcaGhoaFx4eIyUnJSMeLy8zMy8vQEBAQEBAQEBAQEBAQEBAAREPDxETERUSEhUUERQRFBoUFhYUGiYaGhwaGiYwIx4eHh4jMCsuJycnLis1NTAwNTVAQD9AQEBAQEBAQEBAQED/wAARCAADAA4DASIAAhEBAxEB/8QAXAABAQEAAAAAAAAAAAAAAAAAAAIEAQEAAAAAAAAAAAAAAAAAAAACEAAAAwYHAQAAAAAAAAAAAAAAERMBAhIyYhQhkaEDIwUVNREBAAAAAAAAAAAAAAAAAAAAAP/aAAwDAQACEQMRAD8A3dkr5e8tfpwuneJITOzIcmQpit037Bw4mnCVNOpAAQv/2Q=='
+      },
+      {
+        name: 'ladies_outerwear',
+        title: 'Ladies Outerwear',
+        image: 'images/ladies_outerwear.jpg',
+        placeholder: 'data:image/jpeg;base64,/9j/4QAYRXhpZgAASUkqAAgAAAAAAAAAAAAAAP/sABFEdWNreQABAAQAAAAeAAD/7gAOQWRvYmUAZMAAAAAB/9sAhAAQCwsLDAsQDAwQFw8NDxcbFBAQFBsfFxcXFxcfHhcaGhoaFx4eIyUnJSMeLy8zMy8vQEBAQEBAQEBAQEBAQEBAAREPDxETERUSEhUUERQRFBoUFhYUGiYaGhwaGiYwIx4eHh4jMCsuJycnLis1NTAwNTVAQD9AQEBAQEBAQEBAQED/wAARCAADAA4DASIAAhEBAxEB/8QAWQABAQAAAAAAAAAAAAAAAAAAAAEBAQEAAAAAAAAAAAAAAAAAAAIDEAABAwMFAQAAAAAAAAAAAAARAAEygRIDIlITMwUVEQEBAAAAAAAAAAAAAAAAAAAAQf/aAAwDAQACEQMRAD8Avqn5meQ0kwk1UyclmLtNj7L4PQoioFf/2Q=='
+      },
+      {
+        name: 'mens_tshirts',
+        title: 'Men\'s T-Shirts',
+        image: 'images/mens_tshirts.jpg',
+        placeholder: 'data:image/jpeg;base64,/9j/4QAYRXhpZgAASUkqAAgAAAAAAAAAAAAAAP/sABFEdWNreQABAAQAAAAeAAD/7gAOQWRvYmUAZMAAAAAB/9sAhAAQCwsLDAsQDAwQFw8NDxcbFBAQFBsfFxcXFxcfHhcaGhoaFx4eIyUnJSMeLy8zMy8vQEBAQEBAQEBAQEBAQEBAAREPDxETERUSEhUUERQRFBoUFhYUGiYaGhwaGiYwIx4eHh4jMCsuJycnLis1NTAwNTVAQD9AQEBAQEBAQEBAQED/wAARCAADAA4DASIAAhEBAxEB/8QAWwABAQEAAAAAAAAAAAAAAAAAAAMEAQEAAAAAAAAAAAAAAAAAAAAAEAABAwEJAAAAAAAAAAAAAAARAAESEyFhodEygjMUBREAAwAAAAAAAAAAAAAAAAAAAEFC/9oADAMBAAIRAxEAPwDb7kupZU1MTGnvOCgxpvzEXTyRElCmf//Z'
+      },
+      {
+        name: 'ladies_tshirts',
+        title: 'Ladies T-Shirts',
+        image: 'images/ladies_tshirts.jpg',
+        placeholder: 'data:image/jpeg;base64,/9j/4QAYRXhpZgAASUkqAAgAAAAAAAAAAAAAAP/sABFEdWNreQABAAQAAAAeAAD/7gAOQWRvYmUAZMAAAAAB/9sAhAAQCwsLDAsQDAwQFw8NDxcbFBAQFBsfFxcXFxcfHhcaGhoaFx4eIyUnJSMeLy8zMy8vQEBAQEBAQEBAQEBAQEBAAREPDxETERUSEhUUERQRFBoUFhYUGiYaGhwaGiYwIx4eHh4jMCsuJycnLis1NTAwNTVAQD9AQEBAQEBAQEBAQED/wAARCAADAA4DASIAAhEBAxEB/8QAXwABAQEAAAAAAAAAAAAAAAAAAAMFAQEBAAAAAAAAAAAAAAAAAAABAhAAAQIDCQAAAAAAAAAAAAAAEQABITETYZECEjJCAzMVEQACAwAAAAAAAAAAAAAAAAAAATFBgf/aAAwDAQACEQMRAD8AzeADAZiFc5J7BC9Scek3VrtooilSNaf/2Q=='
+      }
+    ]
+  });
+}
+initCategories();
+
+export function changeOffline(offline) {
+  return (dispatch, getState) => {
+    dispatch({
+      type: '_offlineChanged',
+      offline
+    });
+    if (!offline) {
+      tryReconnect()(dispatch, getState);
+    }
+  };
+}
+
+export function tryReconnect() {
+  return (dispatch, getState) => {
+    const state = getState();
+    const category = findCategory(state.categories, state.categoryName);
+    loadCategory(category, dispatch);
+  };
+}
+
+export function loadCategory(category, dispatch) {
+  if (category) {
+    const categoryName = category.name;
+    if (!category.items) {
+      fetch(`data/${categoryName}.json`)
+      .then(res => res.json())
+      .then(data => dispatch({
+        type: '_categoryItemsChanged',
+        categoryName,
+        data
+      }))
+      .catch(() => dispatch({
+        type: '_failureChanged',
+        failure: true
+      }));
+    }
+  }
+}
