+++ src/shop-redux-store.js
@@ -3,39 +3,9 @@ import applyMiddleware from '../node_modules/@0xcda7a/redux-es6/es/applyMiddlewa
 import origCompose from '../node_modules/@0xcda7a/redux-es6/es/compose.js';
 import thunk from '../node_modules/redux-thunk/es/index.js';
 
-import { findCategoryIndex } from './shop-redux-helpers.js';
-
 const compose = window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || origCompose;
 
-const reducers = {
-  // Response from fetch for categories data.
-  _categoryItemsChanged(state, action) {
-    const categories = state.categories.slice(0);
-    const categoryIndex = findCategoryIndex(categories, action.categoryName);
-    categories[categoryIndex] = {...categories[categoryIndex], items: action.data};
-    return {
-      ...state,
-      categories,
-      failure: false
-    };
-  },
-  // Network error (set to true by unsucessful fetch).
-  _failureChanged(state, action) {
-    const failure = action.failure;
-    return {
-      ...state,
-      failure
-    };
-  },
-  // Opposite of navigator.onLine (updated by shop-app).
-  _offlineChanged(state, action) {
-    const offline = action.offline;
-    return {
-      ...state,
-      offline
-    };
-  }
-};
+const reducers = {};
 
 export function installReducers(r) {
   Object.assign(reducers, r);
@@ -46,36 +16,4 @@ export const store = createStore(
     const r = reducers[action.type];
     return r ? r(state, action) : state;
   },
-  getInitialState(),
   compose(applyMiddleware(thunk)));
-
-function getInitialState() {
-  return {
-    categories: [
-      {
-        name: 'mens_outerwear',
-        title: 'Men\'s Outerwear',
-        image: 'images/mens_outerwear.jpg',
-        placeholder: 'data:image/jpeg;base64,/9j/4QAYRXhpZgAASUkqAAgAAAAAAAAAAAAAAP/sABFEdWNreQABAAQAAAAeAAD/7gAOQWRvYmUAZMAAAAAB/9sAhAAQCwsLDAsQDAwQFw8NDxcbFBAQFBsfFxcXFxcfHhcaGhoaFx4eIyUnJSMeLy8zMy8vQEBAQEBAQEBAQEBAQEBAAREPDxETERUSEhUUERQRFBoUFhYUGiYaGhwaGiYwIx4eHh4jMCsuJycnLis1NTAwNTVAQD9AQEBAQEBAQEBAQED/wAARCAADAA4DASIAAhEBAxEB/8QAXAABAQEAAAAAAAAAAAAAAAAAAAIEAQEAAAAAAAAAAAAAAAAAAAACEAAAAwYHAQAAAAAAAAAAAAAAERMBAhIyYhQhkaEDIwUVNREBAAAAAAAAAAAAAAAAAAAAAP/aAAwDAQACEQMRAD8A3dkr5e8tfpwuneJITOzIcmQpit037Bw4mnCVNOpAAQv/2Q=='
-      },
-      {
-        name: 'ladies_outerwear',
-        title: 'Ladies Outerwear',
-        image: 'images/ladies_outerwear.jpg',
-        placeholder: 'data:image/jpeg;base64,/9j/4QAYRXhpZgAASUkqAAgAAAAAAAAAAAAAAP/sABFEdWNreQABAAQAAAAeAAD/7gAOQWRvYmUAZMAAAAAB/9sAhAAQCwsLDAsQDAwQFw8NDxcbFBAQFBsfFxcXFxcfHhcaGhoaFx4eIyUnJSMeLy8zMy8vQEBAQEBAQEBAQEBAQEBAAREPDxETERUSEhUUERQRFBoUFhYUGiYaGhwaGiYwIx4eHh4jMCsuJycnLis1NTAwNTVAQD9AQEBAQEBAQEBAQED/wAARCAADAA4DASIAAhEBAxEB/8QAWQABAQAAAAAAAAAAAAAAAAAAAAEBAQEAAAAAAAAAAAAAAAAAAAIDEAABAwMFAQAAAAAAAAAAAAARAAEygRIDIlITMwUVEQEBAAAAAAAAAAAAAAAAAAAAQf/aAAwDAQACEQMRAD8Avqn5meQ0kwk1UyclmLtNj7L4PQoioFf/2Q=='
-      },
-      {
-        name: 'mens_tshirts',
-        title: 'Men\'s T-Shirts',
-        image: 'images/mens_tshirts.jpg',
-        placeholder: 'data:image/jpeg;base64,/9j/4QAYRXhpZgAASUkqAAgAAAAAAAAAAAAAAP/sABFEdWNreQABAAQAAAAeAAD/7gAOQWRvYmUAZMAAAAAB/9sAhAAQCwsLDAsQDAwQFw8NDxcbFBAQFBsfFxcXFxcfHhcaGhoaFx4eIyUnJSMeLy8zMy8vQEBAQEBAQEBAQEBAQEBAAREPDxETERUSEhUUERQRFBoUFhYUGiYaGhwaGiYwIx4eHh4jMCsuJycnLis1NTAwNTVAQD9AQEBAQEBAQEBAQED/wAARCAADAA4DASIAAhEBAxEB/8QAWwABAQEAAAAAAAAAAAAAAAAAAAMEAQEAAAAAAAAAAAAAAAAAAAAAEAABAwEJAAAAAAAAAAAAAAARAAESEyFhodEygjMUBREAAwAAAAAAAAAAAAAAAAAAAEFC/9oADAMBAAIRAxEAPwDb7kupZU1MTGnvOCgxpvzEXTyRElCmf//Z'
-      },
-      {
-        name: 'ladies_tshirts',
-        title: 'Ladies T-Shirts',
-        image: 'images/ladies_tshirts.jpg',
-        placeholder: 'data:image/jpeg;base64,/9j/4QAYRXhpZgAASUkqAAgAAAAAAAAAAAAAAP/sABFEdWNreQABAAQAAAAeAAD/7gAOQWRvYmUAZMAAAAAB/9sAhAAQCwsLDAsQDAwQFw8NDxcbFBAQFBsfFxcXFxcfHhcaGhoaFx4eIyUnJSMeLy8zMy8vQEBAQEBAQEBAQEBAQEBAAREPDxETERUSEhUUERQRFBoUFhYUGiYaGhwaGiYwIx4eHh4jMCsuJycnLis1NTAwNTVAQD9AQEBAQEBAQEBAQED/wAARCAADAA4DASIAAhEBAxEB/8QAXwABAQEAAAAAAAAAAAAAAAAAAAMFAQEBAAAAAAAAAAAAAAAAAAABAhAAAQIDCQAAAAAAAAAAAAAAEQABITETYZECEjJCAzMVEQACAwAAAAAAAAAAAAAAAAAAATFBgf/aAAwDAQACEQMRAD8AzeADAZiFc5J7BC9Scek3VrtooilSNaf/2Q=='
-      }
-    ]
-  };
-}
